+++ lib/butter.dart
@@ -10,5 +10,7 @@ export 'src/base_page_state.dart';
 export 'src/base_page_transition.dart';
 export 'src/base_page_view.dart';
 export 'src/base_routes.dart';
+export 'src/base_stateful_page_view.dart';
+export 'src/base_stateless_page_view.dart';
 export 'src/base_ui_model.dart';
 export 'src/page_arguments.dart';