+++ lib/src/base_stateless_page_view.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/widgets.dart';
+
+import 'base_page_view.dart';
+
+abstract class BaseStatelessPageView extends StatelessWidget implements BasePageView {
+
+}