+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.2.0] 
+
+* Split BasePageView into BaseStatefulPageView and BaseStatelessPageView
+* Provided BaseStatefulPageView to support smart page loads and updates
+
 ## [0.1.0] 
 
 * Incremented minor version due to breaking changes